댓글 및 대댓글 추가,삭제 기능

+++ src/main/java/com/pickupluck/ecogging/domain/comment/repository/CommentRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
 
-    List<Comment> findByBoardTypeAndArticleId(BoardType boardType, Integer articleId);
+    List<Comment> findByBoardTypeAndArticleId(BoardType boardType, Long articleId);
 
     List<Comment> findByWriter(User Writer);
 }
